src/modules/accounts/useCases/authenticateUser/AuthenticateUserUseCase.spec.ts

import { AppError } from "@shared/errors/AppError";

import { UserRepositoryInMemory } from "@modules/accounts/repositories/in-memory/UsersRepositoryInMemory";
import { AuthenticateUserUseCase } from "./AuthenticateUserUseCase"
import { CreateUserUseCase } from "../createUser/CreateUserUseCase";
import { ICreateUserDTO } from "@modules/accounts/dtos/ICreateUserDTO";

let authenticateUserUseCase: AuthenticateUserUseCase;
let usersRepositoryInMemory: UserRepositoryInMemory;
let createUserUseCase: CreateUserUseCase;

describe("Authenticate User", () => {
    beforeEach(() => {
        usersRepositoryInMemory = new UserRepositoryInMemory();
        createUserUseCase = new CreateUserUseCase(usersRepositoryInMemory);
        authenticateUserUseCase = new AuthenticateUserUseCase(usersRepositoryInMemory);

    })


    it("should be able to authenticate an user", async () => {
        const user: ICreateUserDTO = {
            name: "user Test",
            email: "user@test.com",
            driver_license: "00123",
            password: "1234"
        }

        await createUserUseCase.execute(user);

        await usersRepositoryInMemory.findByEmail(user.email)



        const result = await authenticateUserUseCase.execute({
            email: user.email,
            password: user.password
        })



        expect(result).toHaveProperty("token");

    });


    it("should not be able to authenticate an nonexistent user", async () => {
        expect(async () => {
            await authenticateUserUseCase.execute({
                email: "false@email.com",
                password: "1234"
            })
        }).rejects.toBeInstanceOf(AppError)
    });

    it("should not be able to authenticate with incorrect password", () => {
        expect(async () => {
            const user: ICreateUserDTO = {
                driver_license: "9999",
                email: "user@user.com",
                password: "1234",
                name: "User test"
            }

            await createUserUseCase.execute(user)

            await authenticateUserUseCase.execute({
                email: user.email,
                password: "2533"
            })
        }).rejects.toBeInstanceOf(AppError);
    })
})